Remove code smell in simple_i2c_controller.c
This does not change behavior but it makes it explicit. Overflow of `len+1` is impossible because of C's integer promotion rules and that `tx_len` is an unsigned int, but the code reads like it could be subject to overflow.
This is consistent with an existing cast in `i2c_simple_read_data_generic`.