`i2c_simple_write_data_generic()` passes data to `memcpy()` unchecked
Taking argument input and feeding it into `memcpy()` without checking to determine how long that data is amounts to a classic buffer overflow vulnerability. This could at the very least cause a crash if used improperly; in the worst case, code can be arbitrarily injected into memory this way. Submitting this PR to begin process of fixing.